Removed code i didn't understand what it does. Added code that i still don't know what it does(kidding... not)... Yes i am kidding.
Gallileo Figaro - magnifico

+++ src/main/java/com/viste/realisticarmortiers/capability/ArmorStorage.java
@@ -45,7 +45,7 @@ public class ArmorStorage implements IStorage<IArmor> {
         
         bigCompoundList.setTag("items", itemTagList);
         bigCompoundList.setTag("potions", potionTagList);
-        bigCompoundList.setFloat("speed", instance.getSpeed());
+        bigCompoundList.setDouble("speed", instance.getSpeed());
         bigCompoundList.setTag("usedPotions", usedPotionTagList);
         
         return bigCompoundList;
