+++ src/main/java/com/viste/realisticarmortiers/capability/IArmor.java
@@ -14,13 +14,13 @@ public interface IArmor {
     public void addPotionEffectList(List<Potion> potions);
     public void removePotionEffect(Potion potion);
     public void removeAllItems();
-    public void setSpeed(float speed);
+    public void setSpeed(double speed);
     public void removeUsedPotion(Potion potion);
     public void addUsedPotion(Potion potion);
     
     public List<Potion> getPotionEffect();
     public List<ItemStack> getItems();
-	public float getSpeed();
+	public double getSpeed();
 	public List<Potion> getUsedPotionEffect();
